+++ app/tms/pricing/templates/write/endpoints/create.py
@@ -5,7 +5,7 @@ from sqlalchemy.orm import Session
 
 from app.user.deps.auth import get_current_user
 from app.db.deps import get_db
-from app.models.shipping_provider import ShippingProvider
+from app.tms.providers.models.shipping_provider import ShippingProvider
 from app.tms.pricing.templates.models.shipping_provider_pricing_template import ShippingProviderPricingTemplate
 from app.tms.permissions import check_config_perm
